move enhanced data table under packages
