+++ components/BiospecimenTable.tsx
@@ -9,8 +9,9 @@ import {
     sortByParentID,
 } from '../lib/dataTableHelpers';
 import { Atlas, Entity } from '../lib/helpers';
-import EnhancedDataTable from './EnhancedDataTable';
 import { GenericAttributeNames } from '../lib/types';
+
+import EnhancedDataTable from '../packages/data-portal-table/src/components/EnhancedDataTable';
 
 interface IBiospecimenTableProps {
     samples: Entity[];
