+++ lib/dataTableHelpers.tsx
@@ -2,8 +2,6 @@ import _ from 'lodash';
 import Tooltip from 'rc-tooltip';
 import * as React from 'react';
 
-import { IEnhancedDataTableColumn } from '../components/EnhancedDataTable';
-import ExpandableText from '../components/ExpandableText';
 import {
     getUniqDependencyIds,
     isNumericalSchemaData,
@@ -11,6 +9,9 @@ import {
     SchemaDataId,
 } from './dataSchemaHelpers';
 import { Atlas, Entity } from './helpers';
+
+import { IEnhancedDataTableColumn } from '../packages/data-portal-table/src/components/EnhancedDataTable';
+import ExpandableText from '../components/ExpandableText';
 
 export function getDefaultDataTableStyle() {
     return {
@@ -61,59 +62,6 @@ export function truncatedTableCell<T>(cellData: T) {
     return value ? (
         <ExpandableText fullText={value} truncateProps={{ lines: 4 }} />
     ) : null;
-}
-
-export function getColumnKey(col: {
-    id?: string | number;
-    name: string | number | React.ReactNode;
-}): string {
-    // if no id exists, just use name for key
-    return (col.id || col.name || '').toString();
-}
-
-export function getColumnVisibilityMap(
-    columns: {
-        name: string;
-        id?: string;
-        visible?: boolean;
-    }[] = []
-): { [columnKey: string]: boolean } {
-    const colVis: { [columnKey: string]: boolean } = {};
-
-    columns.forEach((column) => {
-        // every column is visible by default unless it is flagged otherwise
-        let visible = true;
-
-        if (column.visible !== undefined) {
-            visible = column.visible;
-        }
-
-        colVis[getColumnKey(column)] = visible;
-    });
-
-    return colVis;
-}
-
-export function resolveColumnVisibility(
-    columnVisibilityByColumnDefinition: { [columnKey: string]: boolean },
-    columnVisibility?: { [columnKey: string]: boolean },
-    userSelectedColumnVisibility?: { [columnKey: string]: boolean }
-): { [columnKey: string]: boolean } {
-    let colVis: { [columnKey: string]: boolean };
-
-    // if a custom columnVisibility object is provided use that one
-    if (columnVisibility) {
-        colVis = { ...columnVisibility };
-    } else {
-        colVis = {
-            // resolve visibility by column definition
-            ...columnVisibilityByColumnDefinition,
-            // if exists override with the state from the latest user selection
-            ...(userSelectedColumnVisibility || {}),
-        };
-    }
-
-    return colVis;
 }
 
 function defaultNumericalComparison(
